test: de-duplicate the wire-shape tests onto capture_event
Porting the V0 tests left `event.rs` asserting the capture wire shape, which
`capture_event.rs` already owns — six were straight duplicates (timestamp
normalization, the anon/group/explicit `$process_person_profile` cases, and
`$groups`).

Wire-shape assertions now live only in `capture_event.rs`, next to the builder
they exercise; `event.rs` keeps just its `Event`-level timestamp tests. The four
assertions that were genuinely unique moved across rather than being dropped:
`distinct_id` at the root (with no legacy `$distinct_id` and no duplicate in
properties), auto-generated and caller-overridden `uuid`, and
`$process_person_profile` being absent from *both* options and properties when
unset. The `$lib`-coverage note moved with them.

Net -6 tests per config (375/311/352) — the duplicates, nothing else.
