+++ src/event.rs
@@ -322,127 +322,6 @@ impl Event {
     }
 }
 
-/// Asserts how an [`Event`] lands on the V1 capture wire.
-///
-/// Note on `$lib`/`$lib_version`: V1 never carries them in `properties`. The
-/// SDK sends its identity in the `posthog-sdk-info` header and capture
-/// materializes the properties server-side, so that contract is covered by
-/// `client::capture::tests::build_headers_sdk_info_is_canonical_lib_slash_version`.
-#[cfg(test)]
-pub mod tests {
-    use serde_json::{json, Value};
-    use uuid::Uuid;
-
-    use crate::{capture_event::CaptureEvent, Event};
-
-    /// Serialize an event through the V1 builder, yielding the exact JSON the
-    /// capture request would carry for it.
-    fn build_v1(event: Event) -> Value {
-        serde_json::to_value(CaptureEvent::from_event(&event)).unwrap()
-    }
-
-    /// Assert `$process_person_profile` was lifted to `options` with `expected`
-    /// and no longer appears in `properties` (the sentinel must never reach the
-    /// backend as a property).
-    fn assert_person_profile(json: &Value, expected: bool) {
-        assert_eq!(json["options"]["process_person_profile"], json!(expected));
-        assert!(
-            json["properties"].get("$process_person_profile").is_none(),
-            "$process_person_profile must be lifted out of properties"
-        );
-    }
-
-    #[test]
-    fn serializes_distinct_id_at_root() {
-        let json = build_v1(Event::new("test", "user1"));
-
-        // Canonical field at the event root; the legacy `$distinct_id` spelling
-        // (only tolerated by capture via a serde alias) must not be emitted.
-        assert_eq!(json["distinct_id"], "user1");
-        assert!(json.get("$distinct_id").is_none());
-        assert!(json["properties"].get("distinct_id").is_none());
-    }
-
-    #[test]
-    fn serializes_non_utc_timestamp_as_equivalent_utc_instant() {
-        let mut event = Event::new("test", "user1");
-        event
-            .set_timestamp(
-                chrono::DateTime::parse_from_rfc3339("2023-01-01T10:00:00.123+03:00").unwrap(),
-            )
-            .unwrap();
-
-        let json = build_v1(event);
-        assert_eq!(json["timestamp"], "2023-01-01T07:00:00.123Z");
-    }
-
-    #[test]
-    fn includes_auto_generated_uuid() {
-        let json = build_v1(Event::new("test", "user1"));
-
-        let uuid_str = json["uuid"].as_str().expect("uuid should be present");
-        Uuid::parse_str(uuid_str).expect("uuid should be valid");
-    }
-
-    #[test]
-    fn preserves_overridden_uuid() {
-        let uuid = Uuid::now_v7();
-        let mut event = Event::new("test", "user1");
-        event.set_uuid(uuid);
-
-        let json = build_v1(event);
-        assert_eq!(json["uuid"], uuid.to_string());
-    }
-
-    #[test]
-    fn injects_process_person_profile_for_anon() {
-        assert_person_profile(&build_v1(Event::new_anon("anon_test")), false);
-    }
-
-    #[test]
-    fn injects_process_person_profile_for_group() {
-        let mut event = Event::new("test", "user1");
-        event.add_group("company", "acme");
-        assert_person_profile(&build_v1(event), true);
-    }
-
-    #[test]
-    fn no_process_person_profile_when_unset() {
-        let json = build_v1(Event::new("test", "user1"));
-
-        // Absent everywhere: not defaulted into options, not left in properties.
-        assert!(json["options"].get("process_person_profile").is_none());
-        assert!(json["properties"].get("$process_person_profile").is_none());
-    }
-
-    #[test]
-    fn user_property_wins_over_constructor_default() {
-        let mut event = Event::new_anon("test");
-        // new_anon sets $process_person_profile=false; explicit insert overwrites.
-        event.insert_prop("$process_person_profile", true).unwrap();
-        assert_person_profile(&build_v1(event), true);
-    }
-
-    #[test]
-    fn identified_event_with_explicit_personless() {
-        let mut event = Event::new("test", "user1");
-        event.insert_prop("$process_person_profile", false).unwrap();
-        assert_person_profile(&build_v1(event), false);
-    }
-
-    #[test]
-    fn add_group_overrides_anon_person_profile() {
-        let mut event = Event::new_anon("test");
-        // new_anon sets $process_person_profile=false; add_group forces true.
-        event.add_group("company", "acme");
-
-        let json = build_v1(event);
-        assert_person_profile(&json, true);
-        // $groups stays a property on V1 (it is not a lifted option).
-        assert_eq!(json["properties"]["$groups"], json!({"company": "acme"}));
-    }
-}
-
 #[cfg(test)]
 mod test {
     use std::time::Duration;
